Multiple files was moved into several files
