+++ algorithms/palindrome (копия).json
@@ -1,0 +1,137 @@
+{
+  "alphabet": [ "1", "0", ".", "a", "b", "y", "n", "e"],
+  "blank" : ".",
+  "states" : [ "q0", "q1", "q2", "q3", "q4", "q5", "q6", "q7", "q8", "q9", "q10", "q11", "q12", "q13", "q14", "q15", "q16", "q17"],
+  "initial" : "q15",
+  "finals" : [ "q12" ],
+  "transitions" : {
+    "q0": [
+      { "read" : "0", "to_state": "q2", "write": "b", "action": "RIGHT"},
+      { "read" : "1", "to_state": "q1", "write": "a", "action": "RIGHT"},
+      { "read" : "a", "to_state": "q10", "write": "a", "action": "LEFT"},
+      { "read" : "b", "to_state": "q10", "write": "b", "action": "LEFT"},
+      { "read" : ".", "to_state": "q12", "write": "y", "action": "RIGHT"}
+    ],
+
+    "q1": [
+      { "read" : "0", "to_state": "q1", "write": "0", "action": "RIGHT"},
+      { "read" : "1", "to_state": "q1", "write": "1", "action": "RIGHT"},
+      { "read" : "a", "to_state": "q3", "write": "a", "action": "LEFT"},
+      { "read" : "b", "to_state": "q3", "write": "b", "action": "LEFT"},
+      { "read" : ".", "to_state": "q3", "write": ".", "action": "LEFT"}
+    ],
+
+    "q2": [
+      { "read" : "0", "to_state": "q2", "write": "0", "action": "RIGHT"},
+      { "read" : "1", "to_state": "q2", "write": "1", "action": "RIGHT"},
+      { "read" : "a", "to_state": "q4", "write": "a", "action": "LEFT"},
+      { "read" : "b", "to_state": "q4", "write": "b", "action": "LEFT"},
+      { "read" : ".", "to_state": "q4", "write": ".", "action": "LEFT"}
+    ],
+
+    "q3": [
+      { "read" : "0", "to_state": "q13", "write": "0", "action": "LEFT"},
+      { "read" : "1", "to_state": "q5", "write": "a", "action": "LEFT"},
+      { "read" : "a", "to_state": "q10", "write": "a", "action": "LEFT"},
+      { "read" : "b", "to_state": "q10", "write": "b", "action": "LEFT"}
+    ],
+
+    "q4": [
+      { "read" : "0", "to_state": "q5", "write": "b", "action": "LEFT"},
+      { "read" : "1", "to_state": "q13", "write": "1", "action": "LEFT"},
+      { "read" : "a", "to_state": "q10", "write": "a", "action": "LEFT"},
+      { "read" : "b", "to_state": "q10", "write": "b", "action": "LEFT"}
+    ],
+
+    "q5": [
+      { "read" : "0", "to_state": "q7", "write": "b", "action": "LEFT"},
+      { "read" : "1", "to_state": "q6", "write": "a", "action": "LEFT"},
+      { "read" : "a", "to_state": "q10", "write": "a", "action": "LEFT"},
+      { "read" : "b", "to_state": "q10", "write": "b", "action": "LEFT"}
+    ],
+
+    "q6": [
+      { "read" : "0", "to_state": "q6", "write": "0", "action": "LEFT"},
+      { "read" : "1", "to_state": "q6", "write": "1", "action": "LEFT"},
+      { "read" : "a", "to_state": "q8", "write": "a", "action": "RIGHT"},
+      { "read" : "b", "to_state": "q8", "write": "b", "action": "RIGHT"}
+    ],
+
+    "q7":[
+      { "read" : "0", "to_state": "q7", "write": "0", "action": "LEFT"},
+      { "read" : "1", "to_state": "q7", "write": "1", "action": "LEFT"},
+      { "read" : "a", "to_state": "q9", "write": "a", "action": "RIGHT"},
+      { "read" : "b", "to_state": "q9", "write": "b", "action": "RIGHT"}
+    ],
+
+    "q8": [
+      { "read" : "0", "to_state": "q13", "write": "0", "action": "LEFT"},
+      { "read" : "1", "to_state": "q0", "write": "a", "action": "RIGHT"},
+      { "read" : "a", "to_state": "q10", "write": "a", "action": "LEFT"},
+      { "read" : "b", "to_state": "q10", "write": "b", "action": "LEFT"}
+    ],
+
+    "q9": [
+      { "read" : "0", "to_state": "q0", "write": "b", "action": "RIGHT"},
+      { "read" : "1", "to_state": "q13", "write": "1", "action": "LEFT"},
+      { "read" : "a", "to_state": "q10", "write": "a", "action": "LEFT"},
+      { "read" : "b", "to_state": "q10", "write": "b", "action": "LEFT"}
+    ],
+
+    "q10": [
+      { "read" : "a", "to_state": "q10", "write": "a", "action": "LEFT"},
+      { "read" : "b", "to_state": "q10", "write": "b", "action": "LEFT"},
+      { "read" : ".", "to_state": "q11", "write": ".", "action": "RIGHT"}
+    ],
+
+    "q11": [
+      { "read" : "a", "to_state": "q11", "write": "1", "action": "RIGHT"},
+      { "read" : "b", "to_state": "q11", "write": "0", "action": "RIGHT"},
+      { "read" : ".", "to_state": "q12", "write": "y", "action": "RIGHT"}
+    ],
+
+    "q13": [
+      { "read" : "0", "to_state": "q13", "write": "0", "action": "LEFT"},
+      { "read" : "1", "to_state": "q13", "write": "1", "action": "LEFT"},
+      { "read" : "a", "to_state": "q13", "write": "a", "action": "LEFT"},
+      { "read" : "b", "to_state": "q13", "write": "b", "action": "LEFT"},
+      { "read" : ".", "to_state": "q14", "write": ".", "action": "RIGHT"}
+    ],
+
+    "q14": [
+      { "read" : "0", "to_state": "q14", "write": "0", "action": "RIGHT"},
+      { "read" : "1", "to_state": "q14", "write": "1", "action": "RIGHT"},
+      { "read" : "a", "to_state": "q14", "write": "1", "action": "RIGHT"},
+      { "read" : "b", "to_state": "q14", "write": "0", "action": "RIGHT"},
+      { "read" : ".", "to_state": "q12", "write": "n", "action": "RIGHT"}
+    ],
+
+    "q15": [
+      { "read" : "0", "to_state": "q15", "write": "0", "action": "RIGHT"},
+      { "read" : "1", "to_state": "q15", "write": "1", "action": "RIGHT"},
+      { "read" : "a", "to_state": "q16", "write": "a", "action": "RIGHT"},
+      { "read" : "b", "to_state": "q16", "write": "b", "action": "RIGHT"},
+      { "read" : "y", "to_state": "q16", "write": "y", "action": "RIGHT"},
+      { "read" : "n", "to_state": "q16", "write": "n", "action": "RIGHT"},
+      { "read" : "e", "to_state": "q16", "write": "e", "action": "RIGHT"},
+      { "read" : ".", "to_state": "q17", "write": ".", "action": "LEFT"}
+    ],
+
+    "q16": [
+      { "read" : "0", "to_state": "q16", "write": "0", "action": "RIGHT"},
+      { "read" : "1", "to_state": "q16", "write": "1", "action": "RIGHT"},
+      { "read" : "a", "to_state": "q16", "write": "a", "action": "RIGHT"},
+      { "read" : "b", "to_state": "q16", "write": "b", "action": "RIGHT"},
+      { "read" : "y", "to_state": "q16", "write": "y", "action": "RIGHT"},
+      { "read" : "n", "to_state": "q16", "write": "n", "action": "RIGHT"},
+      { "read" : "e", "to_state": "q16", "write": "e", "action": "RIGHT"},
+      { "read" : ".", "to_state": "q12", "write": "e", "action": "RIGHT"}
+    ],
+
+    "q17": [
+      { "read" : "0", "to_state": "q17", "write": "0", "action": "LEFT"},
+      { "read" : "1", "to_state": "q17", "write": "1", "action": "LEFT"},
+      { "read" : ".", "to_state": "q0", "write": ".", "action": "RIGHT"}
+    ]
+  }
+}
