+++ algorithms/task2.turing
@@ -1,0 +1,212 @@
+{
+    "alphabet": [
+        "1",
+        "0",
+        "a",
+        "b",
+        "y",
+        "n",
+        "e"
+    ],
+    "states": {
+        "q0": {
+            "0": "b R q1",
+            "1": "1 L q13",
+            "comment": "база",
+            "a": "a L q10",
+            "b": "N",
+            "y": "N",
+            "n": "N",
+            "e": "N",
+            "λ": "y N q12"
+        },
+        "q1": {
+            "0": "R",
+            "1": "R",
+            "comment": "поиск 1",
+            "a": "a L q2",
+            "b": "N",
+            "y": "N",
+            "n": "N",
+            "e": "N",
+            "λ": "λ L q2"
+        },
+        "q2": {
+            "0": "0 L q13",
+            "1": "a L q3",
+            "comment": "замена 1",
+            "a": "N",
+            "b": "b L q13",
+            "y": "N",
+            "n": "N",
+            "e": "N",
+            "λ": "N"
+        },
+        "q3": {
+            "0": "0 L q13",
+            "1": "a L q4",
+            "comment": "база (р)",
+            "a": "N",
+            "b": "b L q10",
+            "y": "N",
+            "n": "N",
+            "e": "N",
+            "λ": "N"
+        },
+        "q4": {
+            "0": "L",
+            "1": "L",
+            "comment": "поиск 0",
+            "a": "N",
+            "b": "b R q5",
+            "y": "N",
+            "n": "N",
+            "e": "N",
+            "λ": "N"
+        },
+        "q5": {
+            "0": "b R q0",
+            "1": "1 L q13",
+            "comment": "замена 0",
+            "a": "a L q13",
+            "b": "N",
+            "y": "N",
+            "n": "N",
+            "e": "N",
+            "λ": "N"
+        },
+        "q6": {
+            "0": "N",
+            "1": "N",
+            "comment": "-",
+            "a": "N",
+            "b": "N",
+            "y": "N",
+            "n": "N",
+            "e": "N",
+            "λ": "N"
+        },
+        "q7": {
+            "0": "N",
+            "1": "N",
+            "comment": "-",
+            "a": "N",
+            "b": "N",
+            "y": "N",
+            "n": "N",
+            "e": "N",
+            "λ": "N"
+        },
+        "q8": {
+            "0": "N",
+            "1": "N",
+            "comment": "-",
+            "a": "N",
+            "b": "N",
+            "y": "N",
+            "n": "N",
+            "e": "N",
+            "λ": "N"
+        },
+        "q9": {
+            "0": "N",
+            "1": "N",
+            "comment": "-",
+            "a": "N",
+            "b": "N",
+            "y": "N",
+            "n": "N",
+            "e": "N",
+            "λ": "N"
+        },
+        "q10": {
+            "0": "N",
+            "1": "N",
+            "comment": "прокрутка Y",
+            "a": "L",
+            "b": "L",
+            "y": "N",
+            "n": "N",
+            "e": "N",
+            "λ": "λ R q11"
+        },
+        "q11": {
+            "0": "N",
+            "1": "N",
+            "comment": "расшифровка Y",
+            "a": "1 R q11",
+            "b": "0 R q11",
+            "y": "N",
+            "n": "N",
+            "e": "N",
+            "λ": "y N q12"
+        },
+        "q12": {
+            "0": "!",
+            "1": "!",
+            "comment": "конец",
+            "a": "!",
+            "b": "!",
+            "y": "!",
+            "n": "!",
+            "e": "!",
+            "λ": "!"
+        },
+        "q13": {
+            "0": "L",
+            "1": "L",
+            "comment": "прокрутка N",
+            "a": "L",
+            "b": "L",
+            "y": "N",
+            "n": "N",
+            "e": "N",
+            "λ": "λ R q14"
+        },
+        "q14": {
+            "0": "R",
+            "1": "R",
+            "comment": "расшифровка N",
+            "a": "1 R q14",
+            "b": "0 R q14",
+            "y": "N",
+            "n": "N",
+            "e": "N",
+            "λ": "n N q12"
+        },
+        "q15": {
+            "0": "R",
+            "1": "R",
+            "comment": "check correct",
+            "a": "a R q16",
+            "b": "b R q16",
+            "y": "y R q16",
+            "n": "n R q16",
+            "e": "e R q16",
+            "λ": "λ L q17"
+        },
+        "q16": {
+            "0": "R",
+            "1": "R",
+            "comment": "check KO",
+            "a": "R",
+            "b": "R",
+            "y": "R",
+            "n": "R",
+            "e": "R",
+            "λ": "e N q12"
+        },
+        "q17": {
+            "0": "L",
+            "1": "L",
+            "comment": "check OK",
+            "a": "N",
+            "b": "N",
+            "y": "N",
+            "n": "N",
+            "e": "N",
+            "λ": "λ R q0"
+        }
+    },
+    "word": "n"
+}
