--- turing/MDHandler/MDValidatorFields/Blank.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include <vector>
-#include <string>
-
-void ValidateBlank(char blank, const std::string &input_data_for_tape, const std::vector<char> &alphabet);
