--- turing/MDHandler/MDValidatorFields/Finals.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include <vector>
-#include <string>
-
-void ValidateFinals(const std::vector<std::string> &finals, const std::vector<std::string> &states);
