--- turing/MDHandler/MDValidatorFields/Initial.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include <vector>
-#include <string>
-
-void ValidateInitial(const std::string &initial, const std::vector<std::string> &states);
