--- turing/MDHandler/MDValidatorFields/Name.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <string>
-
-void ValidateName(const std::string &name);
